+++ frontend/src/app/GN2CommonModule/map/map.component.ts
@@ -75,6 +75,8 @@ export class MapComponent implements OnInit {
   public searchFailed = false;
   public locationControl = new FormControl();
   public map: Map;
+  public coordinatesTxt: string; /** Texte pour l'affichage des coordonnées */
+
   constructor(
     private mapService: MapService,
     private _commonService: CommonService,
@@ -177,9 +179,47 @@ export class MapComponent implements OnInit {
       }
     });
 
+    if (AppConfig.MAPCONFIG.DISPLAY_MOUSE_COORDINATES) {
+      this.initDisplayCoordinates()
+    }
+
     setTimeout(() => {
       this.map.invalidateSize();
     }, 50);
+  }
+
+  /** Initialise l'affichage des coordonées au survol de la carte */
+  initDisplayCoordinates() {
+    // au survol de la carte
+    // - on affiche les coordonnées en bas au milieu
+    this.map.on("mousemove", (event: any) => {
+      // 6 decimales
+      const lat = Math.round(event.latlng.lat * 1e6) / 1e6;
+      const lng = Math.round(event.latlng.lng * 1e6) / 1e6;
+      this.coordinatesTxt = `${lat}, ${lng}`;
+    });
+
+    // quand le pointeur de la soouris quitte la carte
+    // - on affiche plus les coordonnées
+    this.map.on("mouseout", () => {
+      this.coordinatesTxt = '';
+    });
+
+    // sur un clik droit
+    // - on copie les coordonées dans le presse papier
+    this.map.on('contextmenu',  (event: any) => {
+      // 6 decimales
+      const lat = Math.round(event.latlng.lat * 1e6) / 1e6;
+      const lng = Math.round(event.latlng.lng * 1e6) / 1e6;
+      this.coordinatesTxt = `${lat}, ${lng}`;
+      navigator.clipboard.writeText(
+        this.coordinatesTxt
+      )
+      this._commonService.regularToaster(
+        'success',
+        `Le point\n(lon: ${lng}, lat: ${lat})\na été copié dans le presse papier`
+      );
+    })
   }
 
   /** Retrocompatibility hack to format map config to the expected format: